+++ packages/af-utils-react-virtual-headless/src/components/Subscription/index.js
@@ -1,30 +1,14 @@
-import { useRef, useReducer, memo } from "react";
-import useSubscription from "hooks/useSubscription";
-import { EVT_ALL, FORCEUPDATE_MODULO } from "constants";
+import { memo, useCallback } from "react";
+import { useSyncExternalStore } from "use-sync-external-store/shim";
+import { EVT_ALL } from "constants";
 
-/*
-    Increment just x + 1 may result in NaN theoretically, so modulo should be used.
-    Smth like ( x + 1 ) % 100000
-    But modulo is slow, so using bitwise & is a way to optimize it.
-    m % n equals m & ( n - 1 ) if n is a power of 2.
-*/
-const inc = x => (x + 1) & FORCEUPDATE_MODULO;
-
-const Subscription = ({ model, children, events = EVT_ALL }) => {
-    const [forceUpdateRenderCounter, forceUpdate] = useReducer(inc, 1);
-    const lastRenderCounterRef = useRef(0);
-    const prevRenderRef = useRef(null);
-
-    useSubscription(model, forceUpdate, events);
-
-    if (lastRenderCounterRef.current !== forceUpdateRenderCounter) {
-        lastRenderCounterRef.current = forceUpdateRenderCounter;
-        prevRenderRef.current = children(model);
-    } else {
-        model._Queue.add(forceUpdate);
-    }
-
-    return prevRenderRef.current;
-};
+const Subscription = ({ model, children, getHash, events = EVT_ALL }) => (
+    useSyncExternalStore(
+        useCallback(listener => model.on(listener, events)[events]),
+        getHash,
+        getHash
+    ),
+    children(model)
+);
 
 export default memo(Subscription);