+++ packages/af-utils-react-virtual-headless/src/hooks/useSubscription/index.js
@@ -1,23 +1,10 @@
-import { useLayoutEffect, useEffect } from "react";
+import { useEffect } from "react";
 import { EMPTY_ARRAY } from "constants";
 
-const useIsomorphicLayoutEffect =
-    typeof window !== "undefined" ? useLayoutEffect : useEffect;
-
-/*
-    Why useLayoutEffect?
-
-    Subscription must happen immediately after render,
-    because model change may happen after render and before useEffect.
-    In this case we would miss model update.
-*/
-const useSubscription = (model, callBack, events, invokeImmediately, deps) =>
-    useIsomorphicLayoutEffect(() => {
-        model.on(callBack, events);
-        if (invokeImmediately) {
-            callBack();
-        }
-        return () => model.off(callBack, events);
-    }, deps || EMPTY_ARRAY);
+const useSubscription = (model, callBack, events, deps) =>
+    useEffect(
+        () => (callBack(), model.on(callBack, events)),
+        deps || EMPTY_ARRAY
+    );
 
 export default useSubscription;