+++ packages/website/components/examples/list/VariableSizeList/index.js
@@ -5,7 +5,6 @@ import {
     List
 } from "@af-utils/react-virtual-list";
 import times from "lodash/times";
-import r from "lodash/random";
 
 const DEFAULT_ROW_COUNT = 20000;
 
@@ -30,7 +29,7 @@ const getEstimatedItemSize = (oldItemSizes, oldScrollSize) =>
 
 const VariableSizeList = () => {
     const [dynamicListRowHeights] = useState(() =>
-        times(DEFAULT_ROW_COUNT, () => r(50, 100))
+        times(DEFAULT_ROW_COUNT, i => 50 + ((i ** 2) & 63))
     );
 
     const model = useVirtual({