+++ packages/af-utils-react-table/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @af-utils/react-table
 
+## 0.0.14
+
+### Patch Changes
+
+-   react 18 final
+-   Updated dependencies
+    -   @af-utils/react-virtual-headless@0.0.11
+    -   @af-utils/styled@0.0.7
+
 ## 0.0.13
 
 ### Patch Changes